requirements.txt
pyrofork
tgcrypto
motor
beautifulsoup4
requests
python-dotenv
aiohttp
aiofiles
